backend/requirements.txt
fastapi==0.109.2
uvicorn==0.27.1
python-multipart==0.0.9
sqlalchemy==2.0.27
asyncpg==0.29.0
pgvector==0.2.4
python-magic==0.4.27
pdfplumber==0.10.3
openai>=1.30.0
httpx>=0.27.0
python-dotenv>=1.0.0 